- Moved WktHelper to util - Added documentation
