+++ tests/CMakeLists.txt
@@ -62,4 +62,4 @@ package_add_test(XmlReader util/XmlReader.cpp)
 package_add_test(Decompressor util/Decompressor.cpp)
 package_add_test(OsmDataFetcher osm/OsmDataFetcher.cpp)
 package_add_test(Osm2ttl osm/Osm2ttl.cpp)
-package_add_test(WktHelper osm/WktHelper.cpp)
+package_add_test(WktHelper util/WktHelper.cpp)
