+++ tests/util/WktHelper.cpp
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with osm-live-updates.  If not, see <https://www.gnu.org/licenses/>.
 
-#include "osm/WktHelper.h"
+#include "util/WktHelper.h"
 #include "gtest/gtest.h"
 
 TEST(WktHelper, createDummyNodeFromPoint) {
